+++ packages/remote-feature-flag-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Uncategorized
+
+- chore(lint): Fix suppressed ESLint errors in `remote-feature-flag-controller` package ([#7428](https://github.com/MetaMask/core/pull/7428))
+
 ### Added
 
 - Add override functionality to remote feature flags ([#7271](https://github.com/MetaMask/core/pull/7271))